requirements.txt
python-docx
pandas
openpyxl
PyPDF2
ollama
